Fix minor style & wording issues in prior notification list
