--- frontend/src/features/PriorNotification/constants.ts
@@ -1,23 +1,0 @@
-import { ALL_SEAFRONT_GROUP } from '@constants/seafront'
-import { RichBoolean } from '@mtes-mct/monitor-ui'
-
-import { ExpectedArrivalPeriod } from './components/PriorNotificationList/constants'
-
-import type { ListFilter } from './components/PriorNotificationList/types'
-
-export const DEFAULT_LIST_FILTER_VALUES: ListFilter = {
-  countryCodes: undefined,
-  expectedArrivalCustomPeriod: undefined,
-  expectedArrivalPeriod: ExpectedArrivalPeriod.IN_LESS_THAN_FOUR_HOURS,
-  fleetSegmentSegments: undefined,
-  gearCodes: undefined,
-  hasOneOrMoreReportings: RichBoolean.BOTH,
-  isLessThanTwelveMetersVessel: RichBoolean.BOTH,
-  isSent: undefined,
-  lastControlPeriod: undefined,
-  portLocodes: undefined,
-  priorNotificationTypes: undefined,
-  seafrontGroup: ALL_SEAFRONT_GROUP,
-  searchQuery: undefined,
-  specyCodes: undefined
-}
